Move getPublishers to Topic class
